Add rest api endpoint to delete documents by filter.
