+++ rest_api/application.py
@@ -1,17 +1,30 @@
 import logging
 
+from pathlib import Path
 import uvicorn
 from fastapi import FastAPI, HTTPException
 from starlette.middleware.cors import CORSMiddleware
 
+from haystack import Pipeline
+from rest_api.config import PIPELINE_YAML_PATH, QUERY_PIPELINE_NAME
 from rest_api.controller.errors.http_error import http_error_handler
-from rest_api.controller.router import router as api_router
 from rest_api.config import ROOT_PATH
+
 
 logging.basicConfig(format="%(asctime)s %(message)s", datefmt="%m/%d/%Y %I:%M:%S %p")
 logger = logging.getLogger(__name__)
 logging.getLogger("elasticsearch").setLevel(logging.WARNING)
 logging.getLogger("haystack").setLevel(logging.INFO)
+
+
+PIPELINE = Pipeline.load_from_yaml(Path(PIPELINE_YAML_PATH), pipeline_name=QUERY_PIPELINE_NAME)
+# TODO make this generic for other pipelines with different naming
+RETRIEVER = PIPELINE.get_node(name="Retriever")
+DOCUMENT_STORE = RETRIEVER.document_store if RETRIEVER else None
+logging.info(f"Loaded pipeline nodes: {PIPELINE.graph.nodes.keys()}")
+
+
+from rest_api.controller.router import router as api_router
 
 
 def get_application() -> FastAPI:
@@ -31,6 +44,7 @@ def get_application() -> FastAPI:
 
 
 app = get_application()
+
 
 logger.info("Open http://127.0.0.1:8000/docs to see Swagger API Documentation.")
 logger.info(
